# Conflicts:
#	arduino/library-arduino.json
#	arduino/library-arduino.properties
#	include/etl/iterator.h
#	include/etl/version.h
#	library.json
#	library.properties
#	test/CMakeLists.txt
#	test/vs2022/etl.vcxproj.filters
#	version.txt

+++ test/test_flat_map.cpp
@@ -41,6 +41,7 @@ SOFTWARE.
 #include "data.h"
 
 #include "etl/flat_map.h"
+#include "etl/limits.h"
 
 namespace
 {
@@ -329,7 +330,7 @@ namespace
     {
       DataDC data;
 
-      CHECK(data.size()== size_t(0UL));
+      CHECK(data.size()== 0UL);
       CHECK(data.empty());
       CHECK(data.capacity()== SIZE);
       CHECK(data.max_size()== SIZE);
@@ -1181,13 +1182,13 @@ namespace
 
       DataNDC data(compare_data.begin(), compare_data.end());
       data.clear();
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
 
       // Do it again to check that clear() didn't screw up the internals.
       data.assign(compare_data.begin(), compare_data.end());
       CHECK(data.size() == compare_data.size());
       data.clear();
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
     }
 
     //*************************************************************************
@@ -1195,13 +1196,13 @@ namespace
     {
       DataInt data(int_data.begin(), int_data.end());
       data.clear();
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
 
       // Do it again to check that clear() didn't screw up the internals.
       data.assign(int_data.begin(), int_data.end());
       CHECK(data.size() == int_data.size());
       data.clear();
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
     }
 
     //*************************************************************************
